Changes for font School House Printed A font
Two changes, both probably need to be revisited:

1. The name-entry code supports only a limited list of
   platform-id, encoding-id en language-id combinations.
   This font used a combination not listed.

2. This font has a kerning table.  The subtable format is 0.
   The code expects format 1.    However the code for format 1
   seems to be correct for format 0, not one according to the
   documentation I saw, both Microsoft and Apple.

So what is wrong with the fixes.

For 1. it should be more generic and maybe be not so strict.

For 2. I don't understand how the code could have worked for fonts
with subtable format 1.   Also the name should reflect it is subformat 0.